Rename FormXField file
